Avoid closing twice senders and receivers Restructure test suite to avoid killing the broker.
