+++ src/test/java/io/vertx/ext/amqp/ReplyTest.java
@@ -147,53 +147,6 @@ public class ReplyTest extends ArtemisTestBase {
   }
 
   @Test(timeout = 20000)
-  public void testConnectionToServerWithoutAnonymousSenderLinkSupport(TestContext context) throws Exception {
-    artemis.stop();
-
-    Async asyncShutdown = context.async();
-    AtomicBoolean linkOpened = new AtomicBoolean();
-
-    MockServer server = new MockServer(vertx, serverConnection -> {
-      serverConnection.openHandler(x -> serverConnection.open());
-      serverConnection.closeHandler(x -> serverConnection.close());
-      serverConnection.sessionOpenHandler(ProtonSession::open);
-      serverConnection.receiverOpenHandler(serverReceiver -> {
-        linkOpened.set(true);
-        serverReceiver.setCondition(ProtonHelper.condition(AmqpError.PRECONDITION_FAILED, "Expected no links"));
-        serverReceiver.close();
-      });
-      serverConnection.senderOpenHandler(serverSender -> {
-        linkOpened.set(true);
-        serverSender.setCondition(ProtonHelper.condition(AmqpError.PRECONDITION_FAILED, "Expected no links"));
-        serverSender.close();
-      });
-    });
-    ((ProtonServerImpl) server.getProtonServer()).setAdvertiseAnonymousRelayCapability(false);
-
-    AmqpClientOptions options = new AmqpClientOptions()
-      .setHost("localhost")
-      .setPort(server.actualPort())
-      .setReplyEnabled(true);
-
-    this.client = AmqpClient.create(vertx, options).connect(res -> {
-      context.assertTrue(res.succeeded(), "Expected start to succeed with not reply manager");
-      context.assertFalse(((AmqpConnectionImpl) res.result()).replyManager().isReplySupported());
-      res.result().close(shutdownRes -> {
-        context.assertTrue(shutdownRes.succeeded());
-        asyncShutdown.complete();
-      });
-    });
-
-    try {
-      asyncShutdown.awaitSuccess();
-    } finally {
-      server.close();
-    }
-
-    context.assertFalse(linkOpened.get());
-  }
-
-  @Test(timeout = 20000)
   public void testReply(TestContext context) {
     Async gotReplyAsync = context.async();
     String destinationName = name.getMethodName();
